Implement Manager Dashboard Feature
