+++ goyang_lidah_jogja/urls.py
@@ -20,4 +20,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('main.urls')),
+    path('event_manager/', include('managerDashboard.urls')),
 ]
