+++ main/models.py
@@ -127,40 +127,6 @@ class Menu(models.Model):
         unique_together = ['restaurant', 'menu_code']
         ordering = ['restaurant', 'menu_code']
 
-class Event(models.Model):
-    title = models.CharField(max_length=255)
-    description = models.TextField()
-    date = models.DateField()
-    time = models.TimeField()
-    location = models.CharField(max_length=255)
-    entrance_fee = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True
-    )
-    image = models.URLField(
-        max_length=500,
-        null=True,
-        blank=True,
-        validators=[URLValidator()]
-    )
-    created_by = models.ForeignKey(
-        UserProfile,
-        on_delete=models.CASCADE,
-        related_name='events',
-        limit_choices_to={'role': 'EVENT_MANAGER'}
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = 'Event'
-        verbose_name_plural = 'Events'
-
 class Tag(models.Model):
     name = models.CharField(max_length=50, unique=True)
     menus = models.ManyToManyField(Menu, related_name='tags')
