+++ managerDashboard/models.py
@@ -1,3 +1,37 @@
 from django.db import models
+from django.core.validators import URLValidator
+from main.models import UserProfile
 
-# Create your models here.
+class Event(models.Model):
+    title = models.CharField(max_length=255)
+    description = models.TextField()
+    date = models.DateField()
+    time = models.TimeField()
+    location = models.CharField(max_length=255)
+    entrance_fee = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
+        null=True,
+        blank=True
+    )
+    image = models.URLField(
+        max_length=500,
+        null=True,
+        blank=True,
+        validators=[URLValidator()]
+    )
+    created_by = models.ForeignKey(
+        UserProfile,
+        on_delete=models.CASCADE,
+        related_name='events',
+        limit_choices_to={'role': 'EVENT_MANAGER'}
+    )
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.title
+
+    class Meta:
+        verbose_name = 'Event'
+        verbose_name_plural = 'Events'
