+++ managerDashboard/views.py
@@ -1,3 +1,48 @@
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
+from django.contrib.auth.decorators import login_required, user_passes_test
+from django.http import JsonResponse
+from .models import Event
+from .forms import EventForm
 
-# Create your views here.
+def is_event_manager(user):
+    return user.is_authenticated and user.profile.role == 'EVENT_MANAGER'
+
+@login_required
+@user_passes_test(is_event_manager)
+def event_manager_dashboard(request):
+    events = Event.objects.filter(created_by=request.user.profile).order_by('-date')
+    return render(request, 'event_manager_dashboard.html', {'events': events})
+
+@login_required
+@user_passes_test(is_event_manager)
+def event_create(request):
+    if request.method == 'POST':
+        form = EventForm(request.POST)
+        if form.is_valid():
+            event = form.save(commit=False)
+            event.created_by = request.user.profile
+            event.save()
+            return JsonResponse({'status': 'success'})
+        return JsonResponse({'status': 'error', 'errors': form.errors})
+    return JsonResponse({'status': 'error', 'message': 'Invalid request method'})
+
+@login_required
+@user_passes_test(is_event_manager)
+def event_update(request, pk):
+    event = get_object_or_404(Event, pk=pk)
+    if request.method == 'POST':
+        form = EventForm(request.POST, instance=event)
+        if form.is_valid():
+            form.save()
+            return JsonResponse({'status': 'success'})
+        return JsonResponse({'status': 'error', 'errors': form.errors})
+    return JsonResponse({'status': 'error', 'message': 'Invalid request method'})
+
+@login_required
+@user_passes_test(is_event_manager)
+def event_delete(request, pk):
+    event = get_object_or_404(Event, pk=pk)
+    if request.method == 'POST':
+        event.delete()
+        return JsonResponse({'status': 'success'})
+    return JsonResponse({'status': 'error', 'message': 'Invalid request method'})
